refactor!: move VCs to core
