+++ packages/core/src/credentialsV1/constants.ts
