+++ packages/core/src/credentialsV1/index.ts
@@ -1,0 +1,16 @@
+/**
+ * Copyright (c) 2018-2023, BOTLabs GmbH.
+ *
+ * This source code is licensed under the BSD 4-Clause "Original" license
+ * found in the LICENSE file in the root directory of this source tree.
+ */
+
+/**
+ * @module @kiltprotocol/vc-export
+ */
+
+export * as KiltCredentialV1 from './KiltCredentialV1.js'
+export * as KiltAttestationProofV1 from './KiltAttestationProofV1.js'
+export * as KiltRevocationStatusV1 from './KiltRevocationStatusV1.js'
+export * as Types from './types.js'
+export * as constants from './constants.js'
