+++ packages/utils/src/SDKErrors.ts
@@ -181,3 +181,9 @@ export class NoProofForStatementError extends SDKError {
 export class CodecMismatchError extends SDKError {}
 
 export class PublicCredentialError extends SDKError {}
+
+export class CredentialMalformedError extends SDKError {}
+
+export class PresentationMalformedError extends SDKError {}
+
+export class ProofMalformedError extends SDKError {}
